requirements.txt
Django==5.2.9
djangorestframework==3.16.1
drf-yasg==1.21.11
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
pytz==2025.2
sqlparse==0.5.4
uritemplate==4.2.0
pyyaml==6.0.3